moved multi-tenancy examples out into separate dir and added readme
